refactor: 여러 feature가 쓰는 쿼리를 shared/apis로 이동
