+++ src/pages/banner/apis/queries/useBannerDetailQuery.ts
