+++ src/pages/home/analytics/useProductShopAnalytics.ts
@@ -24,12 +24,13 @@ import {
   type ProductTabController,
 } from '@pages/home/hooks/useProductTabController';
 import type { SelectedProduct } from '@pages/home/types/productTab';
-import { useRecentFloorPlanQuery } from '@pages/imageSetup/apis/queries/useRecentFloorPlanQuery';
 
 import { GA_EVENTS } from '@shared/analytics/events';
 import { useAnalyticsPageView } from '@shared/analytics/hooks';
 import { SCREEN_NAME } from '@shared/analytics/screenNames';
 import { resolveShopImageEntryRoute } from '@shared/analytics/utils/shop/resolveShopEntryRoute';
+
+import { useRecentFloorPlanQuery } from '@apis/queries/useRecentFloorPlanQuery';
 
 interface UseProductShopAnalyticsOptions {
   productCountViewed: number;
