+++ src/pages/imageSetup/analytics/roomTypeAnalytics.ts
@@ -1,5 +1,4 @@
 import { FILTER_CATEGORIES } from '@pages/imageSetup/constants/floorPlanFilters';
-import type { FloorPlanFilters } from '@pages/imageSetup/types/floorPlan';
 
 import { GA_EVENTS } from '@shared/analytics/events';
 import {
@@ -21,6 +20,7 @@ import {
 import { loginStatusParams } from '@shared/analytics/utils/loginStatus';
 import { getPreviousScreenName } from '@shared/analytics/utils/screenName';
 import { toAnalyticsNull } from '@shared/analytics/utils/toAnalyticsNull';
+import type { FloorPlanFilters } from '@shared/types/floorPlan';
 
 import type {
   ExploreHouseTemplateDetailResponse,
