+++ src/pages/imageSetup/apis/queries/useFurnitureCategoriesQuery.ts
@@ -4,9 +4,8 @@ import type { DashboardCategoriesResponse } from '@apis/__generated__/data-contr
 import { HTTPMethod, request } from '@apis/config/request';
 
 import { API_ENDPOINT } from '@constants/apiEndpoints';
+import { STATIC_DATA_QUERY_OPTIONS } from '@constants/cache';
 import { queryKeys } from '@constants/queryKey';
-
-import { STATIC_DATA_QUERY_OPTIONS } from '../../constants/cache';
 
 export const getFurnitureCategories =
   async (): Promise<DashboardCategoriesResponse> => {
