+++ src/pages/imageSetup/apis/queries/useMoodBoardQuery.ts
@@ -8,9 +8,8 @@ import {
 import { HTTPMethod, request } from '@apis/config/request';
 
 import { API_ENDPOINT } from '@constants/apiEndpoints';
+import { STATIC_DATA_QUERY_OPTIONS } from '@constants/cache';
 import { queryKeys } from '@constants/queryKey';
-
-import { STATIC_DATA_QUERY_OPTIONS } from '../../constants/cache';
 
 export const getMoodBoardImage = async (
   limit = MOOD_BOARD_CONSTANTS.DEFAULT_LIMIT
