+++ src/pages/imageSetup/utils/staticDataPrefetch.ts
@@ -1,15 +1,16 @@
 import type { SentryQueryMeta } from '@shared/monitoring/queryMeta';
 import { MONITORING_SCOPE } from '@shared/monitoring/scope';
+import { DEFAULT_FILTERS } from '@shared/types/floorPlan';
 
+import { getHouseTemplates } from '@apis/queries/useHouseTemplatesQuery';
+
+import { STATIC_DATA_QUERY_OPTIONS } from '@constants/cache';
 import { queryKeys } from '@constants/queryKey';
 
 import { getActivities } from '../apis/queries/useActivitiesQuery';
 import { getFurnitureCategories } from '../apis/queries/useFurnitureCategoriesQuery';
-import { getHouseTemplates } from '../apis/queries/useHouseTemplatesQuery';
 import { getMoodBoardImage } from '../apis/queries/useMoodBoardQuery';
-import { STATIC_DATA_QUERY_OPTIONS } from '../constants/cache';
 import { MOOD_BOARD_CONSTANTS } from '../types/apis/interiorStyle';
-import { DEFAULT_FILTERS } from '../types/floorPlan';
 
 import type { QueryClient } from '@tanstack/react-query';
