+++ src/shared/apis/queries/useHouseTemplateDetailQuery.ts
