+++ src/shared/apis/queries/useLandingQuery.ts
