+++ src/shared/types/floorPlan.ts
